(Refact) seperate UploadForm from page
